requirements.txt
xrpl-py
wxPython
requests
toml